+++ src/com/ruegnerlukas/simplemath/vectors/vec3/Vector3f.java
@@ -455,7 +455,7 @@ public class Vector3f implements IVector3 {
 
 	@Override
 	public Vector3f scale(Number xyz) {
-		return this.scale(xyz);
+		return this.scale(xyz.floatValue());
 	}
